+++ src/index.html
@@ -40,19 +40,18 @@
                 </ul>
 
                 <div class="button-container">
-                    <!-- Import Button -->
-                    <label for="importFile" id="importButton" class="button-style">Import JSON</label>
-                    <input type="file" id="importFile" accept=".json" style="display: none" />
 
-                    <!-- Export Button -->
-                    <a href="#" download="feedData.json" id="exportButton" class="button-style">Export JSON</a>
-
+                    <!-- Add Feed -->
                     <button id="button-add-feed" class="button-style">Add Feed</button>
 
-                    <button id="add-feed-button" class="button-style">Add Feed</button>
-                    <button id="upload-button" class="button-style">Upload</button>
-                    <button id="download-button" class="button-style">Download</button>
-                    <button id="save-state-button" class="button-style">Save State</button>
+                    <!-- Import Button -->
+                    <label id="importButton" for="button-state-upload" class="button-style">Import JSON</label>
+                    <input id="button-state-upload" type="file" accept=".json" style="display: none" aria-label="Load File"/>
+
+                    <!-- Export Button -->
+                    <a href="#" download="feedData.json" id="button-state-download" class="button-style">Export JSON</a>
+
+                    
                 </div>
             </div>
